requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
openai==1.58.1
aiosqlite==0.20.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.37.0
google-api-python-client==2.114.0
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
